cwa_classroom/requirements-test.txt
-r requirements.txt
pytest==8.3.4
pytest-django==4.9.0
pytest-playwright==0.6.2
playwright==1.49.1
xlwt==1.3.0
